Use contact-simple in favor of submit-contact, a regular form tag, serve assets statically
